refactor: make updating it's own package, return Result instead
