+++ updates/Cargo.toml
@@ -1,11 +1,10 @@
 [package]
-name = "cirrus_egui"
+name = "cirrus_updates"
 version = "0.1.0"
 edition = "2021"
 
 [dependencies]
-cirrus_theming = { workspace = true }
+reqwest = {version = "0.12.12", features = ["blocking", "json"]}
+semver = "1.0.24"
 
-egui = {workspace = true}
-toml = {workspace = true}
 serde = {workspace = true, features = ["derive"]}
